requirements.txt
tensorflow==2.12.0
streamlit==1.25.0
Pillow
numpy
